Rename the project ID from exsat-credit-staking to esbtc. (#14619)
