+++ projects/helper/bitcoin-book/esbtc.js
