+++ projects/helper/bitcoin-book/index.js
@@ -101,7 +101,7 @@ const prosper = require('./prosper.js')
 const hotcoin = require('./hotcoin.js')
 const orangex = require('./orangex.js')
 const exmo = require('./exmo.js')
-const ecbtc = require('./ecbtc.js')
+const esbtc = require('./esbtc.js')
 
 const p2pb2b = ['39BFtTzZjj6o2s7eewefFQxqM4617VmhEK']
 const teleswap = [
@@ -258,5 +258,5 @@ module.exports = {
   hotcoin,
   orangex,
   exmo,
-  ecbtc
+  esbtc
 }
